Moved forms from model to forms file
